Upgrade Electron to version 1.6.11 for security fixes.
It looks like electron-packager sources the package version for Electron
out of the ./node_modules folder, so we can use yarn to explicitly pin
versions. Hence going back to the "compatible with version" syntax in
package.json and the updated yarn.lock file.

https://github.com/electron-userland/electron-packager/blob/7fcf8b739243b5915fdc783b9a03bec82b1a0c82/infer.js#L37-L63

The v1.7.x line of Electron is still in beta, but 1.6.11 has a few
security fixes around rejecting invalid certs. This upgrade is to pick
those up.