Use symlinks instead of copying files to make macOS zips smaller.
This uses symlinks in the zip packge (the "-y" flag) instead of
keeping multiple copies of the same file. On macOS this means the
Awsaml.app package can be signed and then zipped without invalidating
the signature. It also makes the macOS zip file about 2x smaller.

Testing showed no changes in the Linux zip when building on both
macOS and Linux.