Add basic account filter/search
This adds a very rudimentary search function for recent logins.  It's a
simple case insenstive search.  It doesn't re-order results or anything
fancy like that.  This should make it easier to quickly find accounts if
you have given them helpful aliases.

Currently it searches both the name and url.  Searching the URL may not
make sense, but it didn't hurt in my tests.

GH-118