Pass all accounts to the refresh container
We read in and parse the metadata file from disk, then pass it in
to the refresh container as a property. This is not ideal, but from
what I can see the `Storage` object only contains the current account
at least while in the refresh route. Might loop back around and
see about getting all account metadata in and out of `Storage`.